server/requirements.txt
torch
transformers
pillow
numpy

